# schemas/web-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JeRyu Web Forge API",
    "description": "BFF for the JeRyu Web Forge. Versioned under /api/v1/. See WEB_WORK_CLAUDE.md §35.7 for the canonical route map.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0-alpha"
  },
  "paths": {
...
    "/api/v1/repos/{repo_id}/merge-requests": {
      "get": {
        "tags": [
          "merge_requests"
        ],
        "operationId": "list_merge_requests",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Merge requests list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListMrResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Repository not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "merge_requests"
        ],
        "operationId": "create_merge_request",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMrRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stub OK (Phase 3 placeholder)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyOk"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "502": {
            "description": "Upstream adapter not implemented"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/merge-requests/{iid}": {
      "get": {
        "tags": [
          "merge_requests"
        ],
        "operationId": "get_merge_request",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "in": "path",
            "description": "MR internal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Merge request detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeRequestDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "MR not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "merge_requests"
        ],
        "operationId": "patch_merge_request",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "in": "path",
            "description": "MR internal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMrRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stub OK (Phase 3 placeholder)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyOk"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "502": {
            "description": "Upstream adapter not implemented"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/merge-requests/{iid}/approve": {
      "post": {
        "tags": [
          "merge_requests"
        ],
        "operationId": "approve_merge_request",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "in": "path",
            "description": "MR internal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveMrRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Approval receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveMrResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed / SHA drift"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/merge-requests/{iid}/blockers": {
      "get": {
        "tags": [
          "merge_requests"
        ],
        "operationId": "get_blockers",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "in": "path",
            "description": "MR internal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Merge passport (blockers)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergePassport"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "MR not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/merge-requests/{iid}/checks": {
      "get": {
        "tags": [
          "merge_requests"
        ],
        "operationId": "get_checks",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "in": "path",
            "description": "MR internal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MR checks summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MrChecksResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "MR not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/merge-requests/{iid}/close": {
      "post": {
        "tags": [
          "merge_requests"
        ],
        "operationId": "close_merge_request",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "in": "path",
            "description": "MR internal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyOk"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
...
    "/api/v1/repos/{repo_id}/merge-requests/{iid}/diff": {
      "get": {
        "tags": [
          "merge_requests"
        ],
        "operationId": "get_diff",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "in": "path",
            "description": "MR internal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MR diff",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "MR not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMrRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "target_branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "EmptyOk": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "ListMrResponse": {
        "type": "object",
        "required": [
          "merge_requests"
        ],
        "properties": {
          "merge_requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MergeRequestSummary"
            }
          }
        }
      },
...
      "MergeRequestDetail": {
        "type": "object",
        "description": "Full MR view returned by `GET /api/v1/repos/{repo_id}/merge-requests/{iid}`.\n\nCarries the same posture summary fields plus the Merge Passport verdict\n(per §35.2.4) so the UI's `MergeGatePanel` can render gate-by-gate detail.",
        "required": [
          "summary",
          "merge_passport"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "merge_passport": {
            "$ref": "#/components/schemas/MergePassport"
          },
          "passport_hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "`passport_hash` mirrors `summary.passport_hash` for ergonomics so the\nUI doesn't need to dive into `summary` to display the verdict identity."
          },
          "summary": {
            "$ref": "#/components/schemas/MergeRequestSummary"
          }
        }
      },
...
    },
    "securitySchemes": {
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jeryu_session"
      }
    }
  },
...
}